+++ src/Spd.Presentation.Licensing/ClientApp/src/app/app-routing.module.ts
@@ -1,23 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { AppRoutes } from './app-routes';
 import { LandingComponent } from './landing.component';
-import { BusinessLicenceApplicationRoutes } from './modules/business-licence-application/business-license-application-routes';
-import { ControllingMemberCrcRoutes } from './modules/controlling-member-crc/controlling-member-crc-routes';
-import { PersonalLicenceApplicationRoutes } from './modules/personal-licence-application/personal-licence-application-routes';
 import { AccessDeniedComponent } from './shared/components/access-denied.component';
-
-export class AppRoutes {
-	public static readonly PERSONAL_LICENCE_APPLICATION = PersonalLicenceApplicationRoutes.MODULE_PATH;
-	public static readonly BUSINESS_LICENCE_APPLICATION = BusinessLicenceApplicationRoutes.MODULE_PATH;
-	public static readonly CONTROLLING_MEMBERS_CRC = ControllingMemberCrcRoutes.MODULE_PATH;
-	public static readonly LANDING = '';
-	public static readonly ACCESS_DENIED = 'access-denied';
-	public static readonly INVITATION_DENIED = 'invitation-denied';
-
-	public static path(route: string): string {
-		return `/${route}`;
-	}
-}
 
 const routes: Routes = [
 	{
