+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/personal-licence-application/components/shared/licence-payment-success.component.ts
@@ -3,7 +3,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { PaymentResponse } from '@app/api/models';
 import { PaymentService } from '@app/api/services';
 import { StrictHttpResponse } from '@app/api/strict-http-response';
-import { AppRoutes } from '@app/app-routing.module';
+import { AppRoutes } from '@app/app-routes';
 import { FileUtilService } from '@app/core/services/file-util.service';
 
 @Component({
